chore(slider stories): break out reused logic to shared setup file
